+++ accessories/forecast.js
@@ -4,12 +4,14 @@ const debug = require('debug')('homebridge-weather-plus'),
 
 let Service,
 	Characteristic,
+	CustomService,
 	CustomCharacteristic;
 
-module.exports = function (_Service, _Characteristic, _CustomCharacteristic)
+module.exports = function (_Service, _Characteristic, _CustomService, _CustomCharacteristic)
 {
 	Service = _Service;
 	Characteristic = _Characteristic;
+	CustomService = _CustomService;
 	CustomCharacteristic = _CustomCharacteristic;
 
 	return ForecastWeatherAccessory;
@@ -70,17 +72,18 @@ function ForecastWeatherAccessory(platform, stationIndex, day)
 			// Temperature is an official homekit characteristic
 			if (name === "TemperatureMax")
 			{
-				// Nothing
+				// Fix for negative temperatures, because they are not supported by homekit
+				this.ForecastService.getCharacteristic(Characteristic.CurrentTemperature).props.minValue = -50;
 			}
 			// Use separate services for these characteristics if compatiblity is "home"
 			else if (this.config.compatibility === "home" && compatibility.types.includes(name))
 			{
-				compatibility.createService(this, name, Service, CustomCharacteristic);
+				compatibility.createService(this, name, Service, Characteristic, CustomCharacteristic);
 			}
 			// Use separate services and the temperature service for these characteristics if compatiblity is "both"
 			else if (this.config.compatibility === "both" && compatibility.types.includes(name))
 			{
-				compatibility.createService(this, name, Service, CustomCharacteristic);
+				compatibility.createService(this, name, Service, Characteristic, CustomCharacteristic);
 				if (name === "Humidity")
 				{
 					this.ForecastService.addCharacteristic(Characteristic.CurrentRelativeHumidity);