+++ CHANGELOG.md
@@ -222,3 +222,9 @@
 
 ## 3.2.3
 * Fixed apparent temperature and dew point not responding if lower than 0°C
+
+## 3.2.4
+* Fixed temperature not responding if lower than 0°C since v3.2.3
+* Fixed apparent temperature and dew point not responding in compatibility mode home if lower than 0°C
+* Fixed apparent temperature and dew point are not converted to F in compatibility mode eve
+* Fixed startup error when using compatibility mode eve2 (eve with grouping) and forecasts